requirements.txt
Django
python-decouple
requests
twilio
pandas
openpyxl
gunicorn
psycopg2-binary
whitenoise
dj-database-url
googletrans